+++ common/src/serialize.rs
@@ -4,5 +4,5 @@ pub trait Serialize {
     /// serialized bytes in. Note that the array can be larger
     /// then the actual number of serialized bytes. The number of
     /// bytes is to show what is actually valid.   
-    fn serialize<const LEN: usize>(&self) -> (u8, [u8; LEN]);
+    fn serialize<const LEN: usize>(&self) -> (usize, [u8; LEN]);
 }
